+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Added a new `json` component, which allows building a JSON API entirely in SQL with SQLPage !
    Now creating an api over your database is as simple as `SELECT 'json' AS component, JSON_OBJECT('hello', 'world') AS contents`.
+ - `SELECT` statements that contain only static values are now interpreted directly by SQLPage, and do not result in a database query. This greatly improves the performance of pages that contain many static elements.
 
 ## 0.8.0 (2023-07-17)